@@ -15,12 +15,12 @@ limitations under the License.
 package org.tensorflow.lite.examples.bertqa.ui;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.support.design.chip.Chip;
+import com.google.android.material.chip.Chip;
 import org.tensorflow.lite.examples.bertqa.R;
 
 /** Adapter class to show question suggestion chips. */
