@@ -17,8 +17,8 @@ package org.tensorflow.lite.examples.bertqa.ml;
 import android.content.Context;
 import android.content.res.AssetFileDescriptor;
 import android.content.res.AssetManager;
-import android.support.annotation.WorkerThread;
 import android.util.Log;
+import androidx.annotation.WorkerThread;
 import com.google.common.base.Joiner;
 import java.io.BufferedReader;
 import java.io.FileInputStream;
