@@ -23,6 +23,7 @@ import unittest
 
 import numpy as np
 import tensorflow as tf
+from tensorflow.compat import v1 as tfv1
 
 # pylint: disable=g-bad-import-order
 from tfltransfer import bases
@@ -348,7 +349,7 @@ class ModelCorrectnessTest(unittest.TestCase):
         input_shape=(IMAGE_SIZE, IMAGE_SIZE, 3),
         include_top=False,
         weights='imagenet')
-    tf.keras.experimental.export_saved_model(mobilenet_keras, mobilenet_dir)
+    tfv1.keras.experimental.export_saved_model(mobilenet_keras, mobilenet_dir)
     cls.mobilenet_dir = mobilenet_dir
 
   def setUp(self):
