@@ -52,8 +52,8 @@ public class TFLiteObjectDetectionAPIModel implements Classifier {
   // Only return this many results.
   private static final int NUM_DETECTIONS = 10;
   // Float model
-  private static final float IMAGE_MEAN = 128.0f;
-  private static final float IMAGE_STD = 128.0f;
+  private static final float IMAGE_MEAN = 127.5f;
+  private static final float IMAGE_STD = 127.5f;
   // Number of threads in the java app
   private static final int NUM_THREADS = 4;
   private boolean isModelQuantized;
