@@ -19,8 +19,8 @@ package org.tensorflow.lite.examples.textclassification;
 import android.content.Context;
 import android.content.res.AssetFileDescriptor;
 import android.content.res.AssetManager;
-import android.support.annotation.WorkerThread;
 import android.util.Log;
+import androidx.annotation.WorkerThread;
 import java.io.BufferedReader;
 import java.io.FileInputStream;
 import java.io.IOException;
