@@ -27,6 +27,7 @@ import tensorflow as tf
 from tensorflow_examples.lite.model_maker.core import compat
 from tensorflow_examples.lite.model_maker.core import file_util
 from tensorflow_examples.lite.model_maker.core.task import hub_loader
+from tensorflow_examples.lite.model_maker.core.task import model_util
 
 import tensorflow_hub as hub
 from tensorflow_hub import registry
@@ -71,8 +72,7 @@ def _get_compat_tf_versions(compat_tf_versions=None):
 def get_num_gpus(num_gpus):
   try:
     tot_num_gpus = len(tf.config.experimental.list_physical_devices('GPU'))
-  except TypeError:
-    tf.compat.v1.logging.warning("Couldn't get the number of gpus.")
+  except (tf.errors.NotFoundError, tf.errors.InternalError):
     tot_num_gpus = max(0, num_gpus)
   if num_gpus > tot_num_gpus or num_gpus == -1:
     num_gpus = tot_num_gpus
@@ -148,6 +148,7 @@ class AverageWordVecModelSpec(object):
   compat_tf_versions = _get_compat_tf_versions(2)
   need_gen_vocab = True
   default_training_epochs = 2
+  default_batch_size = 32
   convert_from_saved_model_tf2 = False
 
   def __init__(self,
@@ -203,15 +204,11 @@ class AverageWordVecModelSpec(object):
       writer.write(tf_example.SerializeToString())
     writer.close()
 
-  def run_classifier(self, train_input_fn, validation_input_fn, epochs,
-                     steps_per_epoch, validation_steps, num_classes):
-    """Creates classifier and runs the classifier training."""
-    if epochs is None:
-      epochs = self.default_training_epochs
-
+  def create_model(self, num_classes, optimizer='rmsprop'):
+    """Creates the keras model."""
     # Gets a classifier model.
     model = tf.keras.Sequential([
-        tf.keras.layers.InputLayer(input_shape=[self.seq_len]),
+        tf.keras.layers.InputLayer(input_shape=[self.seq_len], dtype=tf.int32),
         tf.keras.layers.Embedding(
             len(self.vocab), self.wordvec_dim, input_length=self.seq_len),
         tf.keras.layers.GlobalAveragePooling1D(),
@@ -220,6 +217,20 @@ class AverageWordVecModelSpec(object):
         tf.keras.layers.Dense(num_classes, activation='softmax')
     ])
 
+    model.compile(
+        optimizer=optimizer,
+        loss='sparse_categorical_crossentropy',
+        metrics=['accuracy'])
+
+    return model
+
+  def run_classifier(self, train_input_fn, validation_input_fn, epochs,
+                     steps_per_epoch, validation_steps, num_classes):
+    """Creates classifier and runs the classifier training."""
+    if epochs is None:
+      epochs = self.default_training_epochs
+
+    model = self.create_model(num_classes)
     # Gets training and validation dataset
     train_ds = train_input_fn()
     validation_ds = None
@@ -227,11 +238,6 @@ class AverageWordVecModelSpec(object):
       validation_ds = validation_input_fn()
 
     # Trains the models.
-    model.compile(
-        optimizer=tf.keras.optimizers.Adam(),
-        loss='sparse_categorical_crossentropy',
-        metrics=['accuracy'])
-
     model.fit(
         train_ds,
         epochs=epochs,
@@ -373,6 +379,7 @@ def create_classifier_model(bert_config,
       num_labels,
       kernel_initializer=initializer,
       name='output',
+      activation='softmax',
       dtype=tf.float32)(
           output)
 
@@ -386,6 +393,7 @@ class BertModelSpec(object):
 
   compat_tf_versions = _get_compat_tf_versions(2)
   need_gen_vocab = False
+  default_batch_size = 32
 
   def __init__(
       self,
@@ -459,6 +467,18 @@ class BertModelSpec(object):
     self.convert_from_saved_model_tf2 = convert_from_saved_model_tf2
     self.is_built = False
 
+  def reorder_input_details(self, tflite_input_details):
+    """Reorders the tflite input details to map the order of keras model."""
+    for detail in tflite_input_details:
+      name = detail['name']
+      if 'input_word_ids' in name:
+        input_word_ids_detail = detail
+      elif 'input_mask' in name:
+        input_mask_detail = detail
+      elif 'input_type_ids' in name:
+        input_type_ids_detail = detail
+    return [input_word_ids_detail, input_mask_detail, input_type_ids_detail]
+
   def build(self):
     """Builds the class. Used for lazy initialization."""
     if self.is_built:
@@ -507,21 +527,28 @@ class BertClassifierModelSpec(BertModelSpec):
     classifier_data_lib.file_based_convert_examples_to_features(
         examples, label_names, self.seq_len, self.tokenizer, tfrecord_file)
 
-  def _get_classification_loss_fn(self, num_classes):
-    """Gets the classification loss function."""
+  def create_model(self, num_classes, optimizer='adam'):
+    """Creates the keras model."""
+    bert_model, _ = create_classifier_model(
+        self.bert_config,
+        num_classes,
+        self.seq_len,
+        hub_module_url=self.uri,
+        hub_module_trainable=self.trainable,
+        is_tf2=self.is_tf2)
 
-    def _classification_loss_fn(labels, logits):
-      """Classification loss."""
-      labels = tf.squeeze(labels)
-      log_probs = tf.nn.log_softmax(logits, axis=-1)
-      one_hot_labels = tf.one_hot(
-          tf.cast(labels, dtype=tf.int32), depth=num_classes, dtype=tf.float32)
-      per_example_loss = -tf.reduce_sum(
-          tf.cast(one_hot_labels, dtype=tf.float32) * log_probs, axis=-1)
-      loss = tf.reduce_mean(per_example_loss)
-      return loss
+    # Defines evaluation metrics function, which will create metrics in the
+    # correct device and strategy scope.
+    def metric_fn():
+      return tf.keras.metrics.SparseCategoricalAccuracy(
+          'test_accuracy', dtype=tf.float32)
 
-    return _classification_loss_fn
+    bert_model.compile(
+        optimizer=optimizer,
+        loss=tf.keras.losses.SparseCategoricalCrossentropy(),
+        metrics=[metric_fn()])
+
+    return bert_model
 
   def run_classifier(self, train_input_fn, validation_input_fn, epochs,
                      steps_per_epoch, validation_steps, num_classes):
@@ -530,38 +557,16 @@ class BertClassifierModelSpec(BertModelSpec):
     warmup_steps = int(epochs * steps_per_epoch * 0.1)
     initial_lr = self.learning_rate
 
-    def _get_classifier_model():
-      """Gets a classifier model."""
-      classifier_model, core_model = create_classifier_model(
-          self.bert_config,
-          num_classes,
-          self.seq_len,
-          hub_module_url=self.uri,
-          hub_module_trainable=self.trainable,
-          is_tf2=self.is_tf2)
-
-      classifier_model.optimizer = optimization.create_optimizer(
-          initial_lr, steps_per_epoch * epochs, warmup_steps)
-      return classifier_model, core_model
-
-    loss_fn = self._get_classification_loss_fn(num_classes)
-
-    # Defines evaluation metrics function, which will create metrics in the
-    # correct device and strategy scope.
-    def metric_fn():
-      return tf.keras.metrics.SparseCategoricalAccuracy(
-          'test_accuracy', dtype=tf.float32)
-
     with distribution_utils.get_strategy_scope(self.strategy):
       training_dataset = train_input_fn()
       evaluation_dataset = None
       if validation_input_fn is not None:
         evaluation_dataset = validation_input_fn()
-      bert_model, _ = _get_classifier_model()
-      optimizer = bert_model.optimizer
 
-      bert_model.compile(
-          optimizer=optimizer, loss=loss_fn, metrics=[metric_fn()])
+      optimizer = optimization.create_optimizer(initial_lr,
+                                                steps_per_epoch * epochs,
+                                                warmup_steps)
+      bert_model = self.create_model(num_classes, optimizer)
 
     summary_dir = os.path.join(self.model_dir, 'summaries')
     summary_callback = tf.keras.callbacks.TensorBoard(summary_dir)
@@ -661,6 +666,55 @@ def create_qa_model(bert_config,
       network=bert_encoder, initializer=initializer), bert_encoder
 
 
+def create_qa_model_from_squad(max_seq_length,
+                               hub_module_url,
+                               hub_module_trainable=True,
+                               is_tf2=False):
+  """Creates QA model the initialized from the model retrained on Squad dataset.
+
+  Args:
+    max_seq_length: integer, the maximum input sequence length.
+    hub_module_url: TF-Hub path/url to Bert module that's retrained on Squad
+      dataset.
+    hub_module_trainable: True to finetune layers in the hub module.
+    is_tf2: boolean, whether the hub module is in TensorFlow 2.x format.
+
+  Returns:
+    Keras model that outputs start logits and end logits.
+  """
+  if is_tf2:
+    raise ValueError('Only supports to load TensorFlow 1.x hub module.')
+
+  input_word_ids = tf.keras.layers.Input(
+      shape=(max_seq_length,), dtype=tf.int32, name='input_word_ids')
+  input_mask = tf.keras.layers.Input(
+      shape=(max_seq_length,), dtype=tf.int32, name='input_mask')
+  input_type_ids = tf.keras.layers.Input(
+      shape=(max_seq_length,), dtype=tf.int32, name='input_type_ids')
+
+  squad_bert = hub_loader.HubKerasLayerV1V2(
+      hub_module_url,
+      signature='squad',
+      signature_outputs_as_dict=True,
+      trainable=hub_module_trainable)
+
+  outputs = squad_bert({
+      'input_ids': input_word_ids,
+      'input_mask': input_mask,
+      'segment_ids': input_type_ids
+  })
+  start_logits = tf.keras.layers.Lambda(
+      tf.identity, name='start_positions')(
+          outputs['start_logits'])
+  end_logits = tf.keras.layers.Lambda(
+      tf.identity, name='end_positions')(
+          outputs['end_logits'])
+
+  return tf.keras.Model(
+      inputs=[input_word_ids, input_mask, input_type_ids],
+      outputs=[start_logits, end_logits])
+
+
 class BertQAModelSpec(BertModelSpec):
   """A specification of BERT model for question answering."""
 
@@ -680,7 +734,10 @@ class BertQAModelSpec(BertModelSpec):
       trainable=True,
       predict_batch_size=8,
       do_lower_case=True,
-      is_tf2=True):
+      is_tf2=True,
+      convert_from_saved_model_tf2=False,
+      tflite_output_name=None,
+      init_from_squad_model=False):
     """Initialze an instance with model paramaters.
 
     Args:
@@ -708,14 +765,27 @@ class BertQAModelSpec(BertModelSpec):
       do_lower_case: boolean, whether to lower case the input text. Should be
         True for uncased models and False for cased models.
       is_tf2: boolean, whether the hub module is in TensorFlow 2.x format.
+      convert_from_saved_model_tf2: Convert to TFLite from saved_model in TF
+        2.x.
+      tflite_output_name: Dict, output names for the TFLite model.
+      init_from_squad_model: boolean, whether to initialize from the model that
+        is already retrained on Squad 1.1.
     """
-    super(BertQAModelSpec, self).__init__(uri, model_dir, seq_len, dropout_rate,
-                                          initializer_range, learning_rate,
-                                          distribution_strategy, num_gpus, tpu,
-                                          trainable, do_lower_case, is_tf2)
+    super(BertQAModelSpec,
+          self).__init__(uri, model_dir, seq_len, dropout_rate,
+                         initializer_range, learning_rate,
+                         distribution_strategy, num_gpus, tpu, trainable,
+                         do_lower_case, is_tf2, convert_from_saved_model_tf2)
     self.query_len = query_len
     self.doc_stride = doc_stride
     self.predict_batch_size = predict_batch_size
+    if tflite_output_name is None:
+      tflite_output_name = {
+          'start_logits': 'Identity_1',
+          'end_logits': 'Identity'
+      }
+    self.tflite_output_name = tflite_output_name
+    self.init_from_squad_model = init_from_squad_model
 
   def get_name_to_features(self, is_training):
     """Gets the dictionary describing the features."""
@@ -773,23 +843,23 @@ class BertQAModelSpec(BertModelSpec):
         output_fn=output_fn,
         batch_size=batch_size)
 
-  def train(self, train_input_fn, epochs, steps_per_epoch):
-    """Run bert QA training."""
-    warmup_steps = int(epochs * steps_per_epoch * 0.1)
-
-    def _get_qa_model():
-      """Get QA model and optimizer."""
-      qa_model, core_model = create_qa_model(
+  def create_model(self):
+    """Creates the model for qa task."""
+    if self.init_from_squad_model:
+      return create_qa_model_from_squad(self.seq_len, self.uri, self.trainable,
+                                        self.is_tf2)
+    else:
+      qa_model, _ = create_qa_model(
           self.bert_config,
           self.seq_len,
           hub_module_url=self.uri,
           hub_module_trainable=self.trainable,
           is_tf2=self.is_tf2)
-      optimizer = optimization.create_optimizer(self.learning_rate,
-                                                steps_per_epoch * epochs,
-                                                warmup_steps)
-      qa_model.optimizer = optimizer
-      return qa_model, core_model
+      return qa_model
+
+  def train(self, train_input_fn, epochs, steps_per_epoch):
+    """Run bert QA training."""
+    warmup_steps = int(epochs * steps_per_epoch * 0.1)
 
     def _loss_fn(positions, logits):
       """Get losss function for QA model."""
@@ -799,8 +869,10 @@ class BertQAModelSpec(BertModelSpec):
 
     with distribution_utils.get_strategy_scope(self.strategy):
       training_dataset = train_input_fn()
-      bert_model, _ = _get_qa_model()
-      optimizer = bert_model.optimizer
+      bert_model = self.create_model()
+      optimizer = optimization.create_optimizer(self.learning_rate,
+                                                steps_per_epoch * epochs,
+                                                warmup_steps)
 
       bert_model.compile(
           optimizer=optimizer, loss=_loss_fn, loss_weights=[0.5, 0.5])
@@ -810,6 +882,11 @@ class BertQAModelSpec(BertModelSpec):
     checkpoint_path = os.path.join(self.model_dir, 'checkpoint')
     checkpoint_callback = tf.keras.callbacks.ModelCheckpoint(
         checkpoint_path, save_weights_only=True)
+
+    if not bert_model.trainable_variables:
+      tf.compat.v1.logging.warning(
+          'Trainable variables in the model are empty.')
+      return bert_model
 
     bert_model.fit(
         x=training_dataset,
@@ -877,13 +954,46 @@ class BertQAModelSpec(BertModelSpec):
     else:
       return self._predict_without_distribute_strategy(model, input_fn)
 
-  def evaluate(self, model, input_fn, num_steps, eval_examples, eval_features,
-               predict_file, version_2_with_negative, max_answer_length,
-               null_score_diff_threshold, verbose_logging, output_dir):
+  def reorder_output_details(self, tflite_output_details):
+    """Reorders the tflite output details to map the order of keras model."""
+    for detail in tflite_output_details:
+      name = detail['name']
+      if self.tflite_output_name['start_logits'] == name:
+        start_logits_detail = detail
+      if self.tflite_output_name['end_logits'] == name:
+        end_logits_detail = detail
+    return (start_logits_detail, end_logits_detail)
+
+  def predict_tflite(self, tflite_filepath, input_fn):
+    """Predicts the `input_fn` dataset for TFLite model in `tflite_filepath`."""
+    ds = input_fn()
+    all_results = []
+    lite_runner = model_util.LiteRunner(tflite_filepath,
+                                        self.reorder_input_details,
+                                        self.reorder_output_details)
+    for features, _ in ds:
+      outputs = lite_runner.run(features)
+      for unique_id, start_logits, end_logits in zip(features['unique_ids'],
+                                                     outputs[0], outputs[1]):
+        raw_result = run_squad_helper.RawResult(
+            unique_id=unique_id.numpy(),
+            start_logits=start_logits.tolist(),
+            end_logits=end_logits.tolist())
+        all_results.append(raw_result)
+        if len(all_results) % 100 == 0:
+          tf.compat.v1.logging.info('Made predictions for %d records.',
+                                    len(all_results))
+    return all_results
+
+  def evaluate(self, model, tflite_filepath, input_fn, num_steps, eval_examples,
+               eval_features, predict_file, version_2_with_negative,
+               max_answer_length, null_score_diff_threshold, verbose_logging,
+               output_dir):
     """Evaluate QA model.
 
     Args:
-      model: The model to be evaluated.
+      model: The keras model to be evaluated.
+      tflite_filepath: File path to the TFLite model.
       input_fn: Function that returns a tf.data.Dataset used for evaluation.
       num_steps: Number of steps to evaluate the model.
       eval_examples: List of `squad_lib.SquadExample` for evaluation data.
@@ -906,7 +1016,17 @@ class BertQAModelSpec(BertModelSpec):
     Returns:
       A dict contains two metrics: Exact match rate and F1 score.
     """
-    all_results = self.predict(model, input_fn, num_steps)
+    if model is not None and tflite_filepath is not None:
+      raise ValueError('Exactly one of the paramaters `model` and '
+                       '`tflite_filepath` should be set.')
+    elif model is None and tflite_filepath is None:
+      raise ValueError('At least one of the parameters `model` and '
+                       '`tflite_filepath` are None.')
+
+    if tflite_filepath is not None:
+      all_results = self.predict_tflite(tflite_filepath, input_fn)
+    else:
+      all_results = self.predict(model, input_fn, num_steps)
 
     all_predictions, all_nbest_json, scores_diff_json = (
         squad_lib.postprocess_output(
@@ -940,7 +1060,32 @@ mobilebert_classifier_spec = BertClassifierModelSpec(
     is_tf2=False,
     distribution_strategy='off',
     convert_from_saved_model_tf2=True)
+mobilebert_classifier_spec.default_batch_size = 48
 
+mobilebert_qa_spec = BertQAModelSpec(
+    uri='https://tfhub.dev/google/mobilebert/uncased_L-24_H-128_B-512_A-4_F-4_OPT/1',
+    is_tf2=False,
+    distribution_strategy='off',
+    convert_from_saved_model_tf2=True,
+    learning_rate=5e-05,
+    tflite_output_name={
+        'start_logits': 'StatefulPartitionedCall:1',
+        'end_logits': 'StatefulPartitionedCall:0'
+    })
+mobilebert_qa_spec.default_batch_size = 48
+
+mobilebert_qa_squad_spec = BertQAModelSpec(
+    uri='https://tfhub.dev/google/mobilebert/uncased_L-24_H-128_B-512_A-4_F-4_OPT/squadv1/1',
+    is_tf2=False,
+    distribution_strategy='off',
+    convert_from_saved_model_tf2=True,
+    learning_rate=5e-05,
+    tflite_output_name={
+        'start_logits': 'StatefulPartitionedCall:1',
+        'end_logits': 'StatefulPartitionedCall:0'
+    },
+    init_from_squad_model=True)
+mobilebert_qa_squad_spec.default_batch_size = 48
 
 # A dict for model specs to make it accessible by string key.
 MODEL_SPECS = {
@@ -955,7 +1100,9 @@ MODEL_SPECS = {
     'bert': BertModelSpec,
     'bert_classifier': BertClassifierModelSpec,
     'bert_qa': BertQAModelSpec,
-    'mobilebert_classifier': mobilebert_classifier_spec
+    'mobilebert_classifier': mobilebert_classifier_spec,
+    'mobilebert_qa': mobilebert_qa_spec,
+    'mobilebert_qa_squad': mobilebert_qa_squad_spec,
 }
 
 # List constants for supported models.
@@ -966,7 +1113,7 @@ IMAGE_CLASSIFICATION_MODELS = [
 TEXT_CLASSIFICATION_MODELS = [
     'bert_classifier', 'average_word_vec', 'mobilebert_classifier'
 ]
-QUESTION_ANSWERING_MODELS = ['bert_qa']
+QUESTION_ANSWERING_MODELS = ['bert_qa', 'mobilebert_qa', 'mobilebert_qa_squad']
 
 
 def get(spec_or_str):
