@@ -84,35 +84,13 @@ class ModelUtilTest(tf.test.TestCase):
                    input_dim,
                    max_input_value=1000,
                    atol=1e-04):
-    with tf.io.gfile.GFile(tflite_model_file, 'rb') as f:
-      tflite_model = f.read()
-
     np.random.seed(0)
     random_input = np.random.uniform(
         low=0, high=max_input_value, size=(1, input_dim)).astype(np.float32)
 
-    # Gets output from keras model.
-    keras_output = keras_model.predict(random_input)
-
-    # Gets output from tflite model.
-    interpreter = tf.lite.Interpreter(model_content=tflite_model)
-    interpreter.allocate_tensors()
-    input_details = interpreter.get_input_details()[0]
-    if input_details['dtype'] != np.float32:
-      # Quantize the input
-      scale, zero_point = input_details['quantization']
-      random_input = random_input / scale + zero_point
-      random_input = random_input.astype(input_details['dtype'])
-    interpreter.set_tensor(input_details['index'], random_input)
-    interpreter.invoke()
-    output_details = interpreter.get_output_details()[0]
-    lite_output = interpreter.get_tensor(output_details['index'])
-    if output_details['dtype'] != np.float32:
-      # Dequantize the output
-      scale, zero_point = output_details['quantization']
-      lite_output = lite_output.astype(np.float32)
-      lite_output = (lite_output - zero_point) * scale
-    self.assertTrue(np.allclose(lite_output, keras_output, atol=atol))
+    self.assertTrue(
+        test_util.is_same_output(
+            tflite_model_file, keras_model, random_input, atol=atol))
 
 
 if __name__ == '__main__':
