@@ -19,6 +19,7 @@ from __future__ import print_function
 import os
 
 from absl.testing import parameterized
+import numpy as np
 import tensorflow as tf
 
 from tensorflow_examples.lite.model_maker.core import compat
@@ -54,20 +55,45 @@ class QuestionAnswerTest(tf.test.TestCase, parameterized.TestCase):
     with self.assertRaisesRegex(ValueError, 'Incompatible versions'):
       _ = ms.BertQAModelSpec(trainable=False)
 
-  @parameterized.parameters(
-      ('1.1'),
-      ('2.0'),
-  )
   @test_util.test_in_tf_2
-  def test_bert_model(self, version):
+  def test_bert_model(self):
+    # Only test squad1.1 since it takes too long time for this.
+    version = '1.1'
     model_spec = ms.BertQAModelSpec(trainable=False, predict_batch_size=1)
     train_data, validation_data = _get_data(model_spec, version)
     model = question_answer.create(
         train_data, model_spec=model_spec, epochs=1, batch_size=1)
     self._test_f1_score(model, validation_data, 0.0)
     self._test_export_vocab(model)
-    self._test_export_to_tflite(model)
+    self._test_export_to_tflite(model, validation_data)
     self._test_export_to_saved_model(model)
+    # Comments this due to Out of Memory Error.
+    # self._test_model_without_training(model_spec, train_data, validation_data)
+
+  def _test_model_without_training(self, model_spec, train_data,
+                                   validation_data):
+    # Test without retraining.
+    model = question_answer.create(
+        train_data, model_spec=model_spec, do_train=False)
+    self._test_f1_score(model, validation_data, 0.0)
+    self._test_export_to_tflite(model, validation_data)
+
+  @parameterized.parameters(
+      ('mobilebert_qa', False),
+      ('mobilebert_qa_squad', True),
+  )
+  @test_util.test_in_tf_2
+  def test_mobilebert_model(self, spec, trainable):
+    # Only test squad1.1 since it takes too long time for this.
+    version = '1.1'
+    model_spec = ms.get(spec)
+    model_spec.trainable = trainable
+    model_spec.predict_batch_size = 1
+    train_data, validation_data = _get_data(model_spec, version)
+    model = question_answer.create(
+        train_data, model_spec=model_spec, epochs=1, batch_size=1)
+    self._test_f1_score(model, validation_data, 0.0)
+    self._test_export_to_tflite(model, validation_data, atol=1e-02)
 
   def _test_f1_score(self, model, validation_data, threshold):
     metric = model.evaluate(validation_data)
@@ -80,12 +106,39 @@ class QuestionAnswerTest(tf.test.TestCase, parameterized.TestCase):
     self.assertTrue(os.path.isfile(vocab_output_file))
     self.assertGreater(os.path.getsize(vocab_output_file), 0)
 
-  def _test_export_to_tflite(self, model):
+  def _test_export_to_tflite(self,
+                             model,
+                             validation_data,
+                             threshold=0.0,
+                             atol=1e-04):
     tflite_output_file = os.path.join(self.get_temp_dir(), 'model.tflite')
     model.export(self.get_temp_dir(), export_format=ExportFormat.TFLITE)
 
     self.assertTrue(os.path.isfile(tflite_output_file))
     self.assertGreater(os.path.getsize(tflite_output_file), 0)
+
+    metric = model.evaluate_tflite(tflite_output_file, validation_data)
+    self.assertGreaterEqual(metric['final_f1'], threshold)
+
+    spec = model.model_spec
+    input_word_ids = np.random.randint(
+        low=0,
+        high=len(spec.tokenizer.vocab),
+        size=(1, spec.seq_len),
+        dtype=np.int32)
+    input_mask = np.random.randint(
+        low=0, high=2, size=(1, spec.seq_len), dtype=np.int32)
+    input_type_ids = np.random.randint(
+        low=0, high=2, size=(1, spec.seq_len), dtype=np.int32)
+    random_inputs = (input_word_ids, input_mask, input_type_ids)
+
+    self.assertTrue(
+        test_util.is_same_output(
+            tflite_output_file,
+            model.model,
+            random_inputs,
+            model.model_spec,
+            atol=atol))
 
   def _test_export_to_saved_model(self, model):
     save_model_output_path = os.path.join(self.get_temp_dir(), 'saved_model')
